+++ NEWS.md
@@ -3,6 +3,10 @@ ClimaLand.jl Release Notes
 
 main
 --------
+
+
+v0.15.1
+--------
 - Add a tutorial of a single site calibration of a perfect model PR[#621](https://github.com/CliMA/ClimaLand.jl/pull/621)
 - Add a longrun simulation for the bucket PR[#807](https://github.com/CliMA/ClimaLand.jl/pull/807)
 - Add ground heat flux to snow-soil model PR[#796](https://github.com/CliMA/ClimaLand.jl/pull/796)